requirements.txt
python-dotenv
openai-agents
requests